+++ CHANGELOG.md
@@ -1,3 +1,2 @@
 ## Change
-- sync 0.15.4
-- fix #101 #102
+- fix compat with old forge